requirements.txt
discord.py
fastapi
uvicorn
aiosqlite
python-dotenv